Resolve "Change measurement unit from long to float/double"

Closes #27

See merge request courses/dit113/2024/group-1/silo-management-project!22

+++ wio/JsonSerializable.h
@@ -7,7 +7,7 @@
 class JsonSerializable
 {
     public:
-        virtual std::string ToJson(long value, std::string measurementTime) {
+        virtual std::string ToJson(double value, std::string measurementTime) {
             return "{\"value\": " 
                 + std::to_string(value) 
                 + ", \"dateTime\": \""
